Tela e Pacote
Adicionado tamanho mínimo de tela e adicionado atualização de cliente antes de deletar um pacote. Adicionado opção de não deletar o pacote padrão nenhum.

+++ controllers/pacoteController.js
@@ -1,11 +1,12 @@
 const express = require("express");
 const router = express.Router();
-const Pacote = require("../models/Pacote")
+const Pacote = require("../models/Pacote");
+const Cliente = require("../models/Cliente");
 
 // LISTAR PACOTES
 router.get("/administrador/pacotes/listar", (req, res) => {
-    Pacote.findAll({force: true}).then(pacotes => {
-        res.render("administrador/pacotes/listar", {pacotes: pacotes});
+    Pacote.findAll({ force: true }).then(pacotes => {
+        res.render("administrador/pacotes/listar", { pacotes: pacotes });
     });
 });
 
@@ -13,26 +14,38 @@ router.get("/administrador/pacotes/cadastro", (req, res) => {
     res.render("administrador/pacotes/cadastro")
 });
 
-
-
 // DELETAR UM PACOTE
 router.post("/pacotes/delete", (req, res) => {
     var id = req.body.id;
-    if(id != undefined){ //SE FOR DIFERENTE DE NULO
-        if(!isNaN(id)){ //SE FOR UM NÚMERO  
-            
-            Pacote.destroy({
-                where: {
-                    id: id
-                }
-            }).then(() => {
-                res.redirect("/administrador/pacotes/listar");
-            })
+    if (id != undefined) { //SE FOR DIFERENTE DE NULO
+        if (!isNaN(id)) { //SE FOR UM NÚMERO  
 
-        }else{
+            if (id != 1) {
+                // ATUALIZAR OS CLIENTES QUE POSSUEM O PACOTE A SER DELETADO
+                Cliente.update({ pacoteId: 1 }, {
+                    where: {
+                        pacoteId: id
+                    }
+                }).then(() => {
+
+                    // DELETE O PACOTE
+                    Pacote.destroy({
+                        where: {
+                            id: id
+                        }
+                    }).then(() => {
+                        res.redirect("/administrador/pacotes/listar");
+                    });
+
+                });
+            }else {
+                res.redirect("/administrador/pacotes/listar");
+            }
+
+        } else {
             res.redirect("/administrador/pacotes/listar");
         }
-    }else{
+    } else {
         res.redirect("/administrador/pacotes/listar");
     }
 });
@@ -42,11 +55,11 @@ router.post("/pacote/salvar", (req, res) => {
 
     Pacote.create({
         nome: req.body.inputNome,
-        descricao:req.body.inputDescricao,
-        taxaDesconto:req.body.inputDesconto
-    }).then(function(){
+        descricao: req.body.inputDescricao,
+        taxaDesconto: req.body.inputDesconto
+    }).then(function () {
         res.redirect("/administrador/pacotes/listar")
-    }).catch(function(erro){
+    }).catch(function (erro) {
         res.redirect("/administrador/pacotes/listar")
     })
 })
@@ -55,14 +68,14 @@ router.post("/pacote/salvar", (req, res) => {
 router.get("/administrador/pacotes/editar/:id", (req, res) => {
     var id = req.params.id;
 
-    if(isNaN(id)){
+    if (isNaN(id)) {
         res.redirect("/administrador/pacotes/listar")
     }
 
     Pacote.findByPk(id).then(pacote => {
-        if(pacote != undefined){
-            res.render("administrador/pacotes/editar", {pacote: pacote});
-        }else {
+        if (pacote != undefined) {
+            res.render("administrador/pacotes/editar", { pacote: pacote });
+        } else {
             res.redirect("/administrador/pacotes/listar")
         }
     }).catch(erro => {
